Move integration test to separate module
This is required when we have circular dependencies we want to test.
